+++ app/api/index.js
@@ -1,0 +1,9 @@
+import * as fileAPI from './fileAPI'
+import * as gitAPI from './gitAPI'
+import * as workspaceAPI from './workspaceAPI'
+
+export default {
+  ...fileAPI,
+  ...gitAPI,
+  ...workspaceAPI,
+}
