+++ app/commands/commandBindings/git.js
@@ -2,7 +2,7 @@
 import store from '../../store'
 const { getState, dispatch: $d } = store
 
-import * as api from '../../api'
+import api from '../../api'
 import * as Git from '../../components/Git/actions'
 import * as Modal from '../../components/Modal/actions'
