+++ app/index.js
@@ -5,7 +5,6 @@ import qs from 'querystring'
 
 import Root from './containers/Root.jsx';
 import './styles/main.styl';
-import {setupWorkspace} from './api'
 import config from './config';
 
 const {spaceKey} = qs.parse(window.location.hash.slice(1));
